Change split_to_tokens signature
as noted in a comment, the ideal return type would be optional<vector>.
To get closer to that, the function returns a bool, denoting whether an error has occured
and whether the "returned" vector has content.
The function now takes a pointer to a vector which it will initialize and put data into.
On error, the vector is cleared.

also updated the example