file descritor handler
bash runs through a command, opens all the redirections in sequential order, dups fds to stdout, then executes the programs and writes in descending order before wrapping up the command. I'm trying to follow this as close as possible so everything runs similarly. Got a working redirection opener kind of.... still needs a lot of work and needs to be cleaned up.