file descriptor handler working prototype
file descritor handler is working for redirection. It makes opens a file according to it token, dups the open fd to one that is greater than 10 at least. then puts that fd on a stack where it can be popped later when needed. Also some accessory functions that allow for easy closing of all the fds and freeing vector memory.