changing exiting and errors in execution.
changed msh_error so that it does not exit the program. Will be making returns that return in error call this function to print minishell: error_message on error without closing the shell. This will mean, many functions need to be changed so that no functions exit the program early.