function returns
went through the execution algo to make sure all the returns are appropriate and appropriately handled. ie freeing when necessary, calling error functions to show errors when necessary. We decided not to kill the shell when errors occur so the return from tree_crawler is not checked. but error messages will show and execution continues.