Large commit.
Re-arranged a lot of stuff and did some renaming. Most of the changes were just to get all the function for builtins in the code and give them all the right arguments. I also moved any function that manages a part of the builtin functionality to one of the builtin source files. Also made the return from builtins cascade back up to prompt_loop so that we can set the return int.