parser : fix segfault
Manually get the line + offset from rlinfo instead of calling ft_readline.
Because prompt_loop handles ft_readline returning null, rlinfo->line is guaranted to be non null.
However it could be empty, in which case ft_readline would call readline again, and potentially return NULL
I could have handled the nullptr in the parser as well, but decided not too, to have only a single point
in the code responsible for displaying the prompt.